Move services out of responseobjects directory
