+++ src/azul/service/responseobjects/cart_item_manager.py
@@ -6,7 +6,7 @@ import uuid
 from azul import config
 from azul.service.responseobjects.dynamo_data_access import DynamoDataAccessor
 from azul.service.responseobjects.elastic_request_builder import ElasticTransformDump as EsTd
-from azul.service.responseobjects.step_function_helper import StepFunctionHelper
+from azul.service.step_function_helper import StepFunctionHelper
 
 
 class CartItemManager:
